add connections controllers

+++ server/src/controllers/connection.controller.js
@@ -4,10 +4,10 @@ import { Connection } from "../models/connection.model.js";
 import { User } from "../models/user.model.js";
 
 const createConnection = asyncHandler(async (req, res) => {
-    const { connectionId } = req.body;
+    const { connectionId } = req.params;
     const userId = req.user._id;
 
-    if (connectionId === userId) {
+    if (connectionId == userId) {
         return res
             .status(400)
             .json({ message: "You cannot connect with yourself" });
@@ -40,6 +40,20 @@ const createConnection = asyncHandler(async (req, res) => {
 const acceptConnection = asyncHandler(async (req, res) => {
     const { connectionId } = req.params;
 
+    const isConnection = await Connection.findById(connectionId);
+
+    if (!isConnection) {
+        return res
+            .status(404)
+            .json({ message: "Connection request not found" });
+    }
+
+    if (!isConnection.connectTo.equals(req.user._id)) {
+        return res.status(403).json({
+            message: "You are not authorized to accept this connection",
+        });
+    }
+
     const connection = await Connection.findByIdAndUpdate(
         connectionId,
         {
@@ -68,6 +82,20 @@ const acceptConnection = asyncHandler(async (req, res) => {
 const rejectConnection = asyncHandler(async (req, res) => {
     const { connectionId } = req.params;
 
+    const isConnection = await Connection.findById(connectionId);
+
+    if (!isConnection) {
+        return res
+            .status(404)
+            .json({ message: "Connection request not found" });
+    }
+
+    if (!isConnection.connectTo.equals(req.user._id)) {
+        return res.status(403).json({
+            message: "You are not authorized to reject this connection",
+        });
+    }
+
     const connection = await Connection.findByIdAndDelete(connectionId);
 
     if (!connection) {
@@ -79,11 +107,7 @@ const rejectConnection = asyncHandler(async (req, res) => {
     return res
         .status(200)
         .json(
-            new ApiResponse(
-                200,
-                connection,
-                "Connection request rejected successfully"
-            )
+            new ApiResponse(200, {}, "Connection request rejected successfully")
         );
 });
 
@@ -91,6 +115,7 @@ const getConnections = asyncHandler(async (req, res) => {
     const userId = req.user._id;
 
     const connections = await Connection.find({
+        isAccepted: true,
         $or: [{ connectFrom: userId }, { connectTo: userId }],
     });
 
@@ -105,4 +130,49 @@ const getConnections = asyncHandler(async (req, res) => {
         );
 });
 
-export { createConnection, acceptConnection, rejectConnection, getConnections };
+const getPendingRequests = asyncHandler(async (req, res) => {
+    const userId = req.user._id;
+
+    const connections = await Connection.find({
+        isAccepted: false,
+        connectFrom: userId,
+    });
+
+    return res
+        .status(200)
+        .json(
+            new ApiResponse(
+                200,
+                connections,
+                "Pending Requests connections fetched successfully"
+            )
+        );
+});
+
+const getPendingApprovals = asyncHandler(async (req, res) => {
+    const userId = req.user._id;
+
+    const connections = await Connection.find({
+        isAccepted: false,
+        connectTo: userId,
+    });
+
+    return res
+        .status(200)
+        .json(
+            new ApiResponse(
+                200,
+                connections,
+                "Pending Approvals connections fetched successfully"
+            )
+        );
+});
+
+export {
+    createConnection,
+    acceptConnection,
+    rejectConnection,
+    getConnections,
+    getPendingRequests,
+    getPendingApprovals,
+};
